Fix compilation errors for federated learning
- Fixed VillagerDialogueAI UUID to String conversion
- Fixed DialogueContext type in personality recording
- Fixed DoubleDQN to use no-args constructor with lazy init
- Added getModelPath() public accessor to ModelPersistence
- Added missing methods to MultiAgentLearning:
  * recordTeamExperience()
  * getTeamCoordinationBonus()
  * getTeamCount()
  * formTeam(List) overload
- Added missing methods to DoubleDQN:
  * selectActionIndex()
  * trainBatch()
- Fixed PrioritizedReplayBuffer lossy conversion warning
- Fixed FTBTeamsIntegration cache clearing

All 15 compilation errors resolved.

+++ src/main/java/com/minecraft/gancity/ml/PrioritizedReplayBuffer.java
@@ -91,7 +91,7 @@ public class PrioritizedReplayBuffer {
                     sampledList.add(pExp);
                     
                     // Calculate importance sampling weight
-                    float probability = Math.pow(pExp.priority, alpha) / totalPriority;
+                    float probability = (float) (Math.pow(pExp.priority, alpha) / totalPriority);
                     float weight = (float) Math.pow(buffer.size() * probability, -beta);
                     weights.add(weight);
                     break;
